requirements.txt
aniso8601==3.0.2
argcomplete==1.9.3
argon2-cffi==19.1.0
Babel==2.6.0
boto3==1.9.123
botocore==1.12.123
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
Django==2.1.7
django-admin-env-notice==0.1.5
django-admin-honeypot==1.1.0
django-cors-headers==2.5.2
django-formtools==2.1
django-otp==0.5.2
django-phonenumber-field==1.3.0
django-ratelimit==2.0.0
django-storages==1.7.1
django-two-factor-auth==1.8.0
docutils==0.14
durationpy==0.5
facebook-sdk==3.1.0
future==0.16.0
google-auth==1.6.3
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
phonenumberslite==8.10.8
pip==19.0.3
placebo==0.9.0
promise==2.2.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==5.1
qrcode==6.1
requests==2.21.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
s3transfer==0.2.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.15.1
wheel==0.33.1
wsgi-request-logger==0.4.6
zappa==0.48.2
